requirements.txt
# Backend
fastapi
uvicorn
requests
python-dotenv

# Google Calendar API
google-api-python-client
google-auth

# LangChain + Gemini
langchain
langchain-google-genai
google-generativeai

# Frontend
streamlit
